Move GlobalStyles to mainstage + delete folder utils
