+++ src/App.js
@@ -1,12 +1,17 @@
 import React from 'react'
 import TripList from './components/TripList'
+import styled from 'styled-components'
 
 function App() {
   return (
-    <div className="App">
+    <AppStyled>
       <TripList />
-    </div>
+    </AppStyled>
   )
 }
 
 export default App
+
+const AppStyled = styled.div`
+  max-width: 768px;
+`
